Move to contract folder
