+++ racun-service/src/main/java/rs/edu/raf/banka/racun/model/contract/TransakcionaStavka.java
@@ -1,4 +1,4 @@
-package rs.edu.raf.banka.racun.model;
+package rs.edu.raf.banka.racun.model.contract;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import rs.edu.raf.banka.racun.enums.HartijaOdVrednostiType;
 import rs.edu.raf.banka.racun.enums.RacunType;
 import rs.edu.raf.banka.racun.enums.TransakcionaStavkaType;
+import rs.edu.raf.banka.racun.model.Valuta;
 
 import javax.persistence.*;
