+++ racun-service/src/main/java/rs/edu/raf/banka/racun/repository/contract/TransakcionaStavkaRepository.java
@@ -1,0 +1,8 @@
+package rs.edu.raf.banka.racun.repository.contract;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import rs.edu.raf.banka.racun.model.contract.TransakcionaStavka;
+
+public interface TransakcionaStavkaRepository extends JpaRepository<TransakcionaStavka, Long> {
+
+}
