+++ racun-service/src/main/java/rs/edu/raf/banka/racun/service/impl/TransakcijaService.java
@@ -24,7 +24,6 @@ import javax.persistence.Query;
 import javax.transaction.Transactional;
 import java.util.Date;
 import java.util.List;
-import java.util.UUID;
 
 @Service
 @Slf4j
