+++ racun-service/src/main/java/rs/edu/raf/banka/racun/service/impl/UgovorService.java
@@ -5,10 +5,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import rs.edu.raf.banka.racun.enums.UgovorStatus;
-import rs.edu.raf.banka.racun.model.TransakcionaStavka;
-import rs.edu.raf.banka.racun.model.Ugovor;
+import rs.edu.raf.banka.racun.model.contract.TransakcionaStavka;
+import rs.edu.raf.banka.racun.model.contract.Ugovor;
 import rs.edu.raf.banka.racun.repository.*;
 import rs.edu.raf.banka.racun.repository.company.CompanyRepository;
+import rs.edu.raf.banka.racun.repository.contract.TransakcionaStavkaRepository;
+import rs.edu.raf.banka.racun.repository.contract.UgovorRepository;
 import rs.edu.raf.banka.racun.requests.TransakcionaStavkaCreateRequest;
 import rs.edu.raf.banka.racun.requests.TransakcionaStavkaUpdateRequest;
 import rs.edu.raf.banka.racun.requests.UgovorCreateRequest;
