+++ racun-service/src/test/java/rs/edu/raf/banka/racun/SredstvaKapitalaServiceTest.java
@@ -6,7 +6,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 import rs.edu.raf.banka.racun.dto.*;
